@@ -1,38 +1,33 @@
 {
 	"name": "vapor-discord-bot",
-	"version": "0.0.1",
-	"description": "A discord bot",
-	"main": "dist/src/index.js",
+	"version": "0.1.0",
+	"private": true,
 	"scripts": {
-		"dev:run": "tsc-watch --noClear -p ./tsconfig.json --onSuccess \"node ./dist/src/index.js\"",
+		"lint": "prettier --check . && eslint src --ext .ts --format=pretty",
+		"format": "prettier --write . && eslint src --ext .ts --fix --format=pretty",
+		"start": "NODE_ENV=production node --require dotenv/config prod/index.js",
+		"start:dev": "nodemon -e ts,json --exec \"npm run build:dev\"",
 		"build": "tsc",
-		"fix": "gts fix",
-		"lint": "gts lint",
-		"clean": "gts clean",
-		"test": "jest"
+		"build:dev": "tsc && node --unhandled-rejections=strict --trace-uncaught --trace-warnings --require dotenv/config prod/index.js"
 	},
-	"author": "Sean DeBarr",
-	"license": "MIT",
-	"devDependencies": {
-		"@types/jest": "27.4.0",
-		"@types/node": "18.11.18",
-		"@types/ws": "8.2.2",
-		"gts": "3.1.0",
-		"jest": "27.4.7",
-		"ts-jest": "27.1.2",
-		"tsc-watch": "4.6.0",
-		"typescript": "4.5.4"
-	},
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/setdebarr/vapor-discord-bot.git"
-	},
-	"bugs": {
-		"url": "https://github.com/setdebarr/vapor-discord-bot/issues"
-	},
-	"homepage": "https://github.com/setdebarr/vapor-discord-bot#readme",
 	"dependencies": {
-		"discord.js": "13.6.0",
-		"dotenv": "10.0.0"
+		"discord.js": "^14.11.0",
+		"dotenv": "^16.0.3",
+		"fs-extra": "^11.1.1"
+	},
+	"devDependencies": {
+		"@types/fs-extra": "^11.0.1",
+		"@types/node": "^18.16.14",
+		"@typescript-eslint/eslint-plugin": "^5.59.6",
+		"@typescript-eslint/parser": "^5.59.6",
+		"eslint": "^8.41.0",
+		"eslint-config-neon": "^0.1.47",
+		"eslint-config-prettier": "^8.8.0",
+		"eslint-formatter-pretty": "^5.0.0",
+		"eslint-plugin-prettier": "^4.2.1",
+		"jest": "^29.5.0",
+		"nodemon": "^2.0.22",
+		"prettier": "^2.8.8",
+		"typescript": "^5.0.4"
 	}
 }
