@@ -1,22 +1,25 @@
-require('dotenv').config();
-import Discord = require('discord.js');
+import * as fs from 'fs-extra';
+import * as path from 'node:path';
 
-const client = new Discord.Client({intents: [Discord.Intents.FLAGS.GUILDS]});
+import { Client, GatewayIntentBits } from 'discord.js';
 
-client.on('ready', async () => {
-	console.log('The bot is ready');
-	client!.user!.setActivity('Plotting Human Demise...');
-});
+(async () => {
+	const client = new Client({ intents: [GatewayIntentBits.Guilds] });
+	const eventsPath = path.join(__dirname, 'events');
+	const eventFiles = await fs.readdir(eventsPath).then((files) => files.filter((file) => file.endsWith('.js')));
 
-client.on('message', async message => {
-	if (message.author.bot) return;
-	console.log(`User: ${message.author.username} | ${message.content}`);
-	if (message.content.indexOf('!') !== 0) return;
+	for (const file of eventFiles) {
+		const event = (await import(path.join(eventsPath, file))).default;
+		client[event.once ? 'once' : 'on'](event.name, async (...args) => event.execute(...args));
+	}
 
-	message.reply(
-		'Thank you for summoning me human. Unfortunately I do not have any commands to use yet. Try again later when I have had time to create some.'
+	void client.login(
+		process.env['NODE_ENV'] === 'production' ? process.env['DISCORD_TOKEN'] : process.env['DISCORD_TOKEN_DEV'],
 	);
+})().catch((error) => {
+	console.log('Something went fatally wrong');
+	console.log(error);
+	console.log(error.message);
+	console.log(error.stack);
+	process.exit(1);
 });
-
-// Login to Discord
-client.login(process.env.BOT_TOKEN);
